src/requirements.txt
cornice==2.4.0
gunicorn==19.7.1
hupper==1.0
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
PasteDeploy==1.5.2
plaster==0.5
plaster-pastedeploy==0.4.1
PyGithub==1.35
Pygments==2.2.0
PyJWT==1.5.3
pyramid==1.9.1
pyramid-debugtoolbar==4.3
pyramid-jinja2==2.7
pyramid-mako==1.0.2
repoze.lru==0.7
simplejson==3.11.1
translationstring==1.3
venusian==1.1.0
WebOb==1.7.3
wheel==0.24.0
zope.deprecation==4.3.0
zope.interface==4.4.2
